+++ src/main/java/com/example/Reservations/model/entity/Location.java
@@ -4,6 +4,7 @@ import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.time.Instant;
 
 @Getter
@@ -20,7 +21,7 @@ public class Location {
 
     private String descricao;
 
-    private double valor_hora;
+    private BigDecimal valor_hora;
 
     private int tempo_minimo;
